図鑑サムネイルも大容量画像を圧縮
